# CLN: fix typo in ctors.SeriesDtypesConstructors setup